Fix mismatched types: widen synapse from_index to u32 for GPU struct
neat-core's `SynapseData.from_index` is `u16`, but `SynapseGpu.from_index`
must be `u32` (WGSL has no u16). Convert explicitly with `u32::from(..)`,
mirroring the existing `u32::from(n.num_synapses)` pattern, fixing the
E0308 build error broken in consumers (stSoftwareAU/NEAT-AI-Examples#604).

Vibe-Coder-Run-Id: vibe-mqqdqjwm-560862

+++ rust_scorer/src/gpu/forward_mse_batched.rs
@@ -225,7 +225,7 @@ pub fn build_batched_network_data(
         for s in &net.synapses {
             synapses.push(SynapseGpu {
                 weight: s.weight,
-                from_index: s.from_index,
+                from_index: u32::from(s.from_index),
             });
         }
